+++ perm_hmm/loss_functions.py
@@ -27,6 +27,7 @@ def log_binary_zero_one(dark_state, bright_state):
 def binary_zero_one(dark_state, bright_state):
     testing_states = torch.tensor([dark_state, bright_state], dtype=int)
 
+
     def _wrapper(state, classification):
-        return zero_one(state, testing_states[classification])
+        return zero_one(state, testing_states[classification.long()])
     return _wrapper
